fix (refs T33506): Fix wrong namespaces after move
